Re-add old normal generation code when no normals are present in the OBJ at all (fixes Linsmeau) Fix normal parsing not respecting the chosen vertex order (fixes Copan) Mark objects with d=0 as fully transparent on loading Allow up to 4 textures to be loaded in parallel on x64 Bind OpenGL context on shader reload Some minor fixes

+++ src/core/StelOpenGLArray.cpp
@@ -215,7 +215,7 @@ bool StelOpenGLArray::load(const StelOBJ* obj, bool useTangents)
 			m_indexBufferType = GL_UNSIGNED_SHORT;
 			m_indexBufferTypeSize = sizeof(GLushort);
 			const StelOBJ::ShortIndexList idxList = obj->getShortIndexList();
-			m_indexBuffer.allocate(idxList.constData(),m_indexBufferTypeSize * idxList.size());
+			m_indexBuffer.allocate(idxList.constData(),static_cast<int>(m_indexBufferTypeSize * idxList.size()));
 		}
 		else
 		{
@@ -231,7 +231,7 @@ bool StelOpenGLArray::load(const StelOBJ* obj, bool useTangents)
 			m_indexBufferTypeSize = sizeof(GLuint);
 
 			const StelOBJ::IndexList& idxList = obj->getIndexList();
-			m_indexBuffer.allocate(idxList.constData(),m_indexBufferTypeSize * idxList.size());
+			m_indexBuffer.allocate(idxList.constData(),static_cast<int>(m_indexBufferTypeSize * idxList.size()));
 		}
 		m_indexCount = obj->getIndexList().size();
 		m_memoryUsage+= m_indexCount * m_indexBufferTypeSize;
